+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.0.8]
 
 ### Fixed
-- Use datetime objects for TIMESTAMP device class sensors instead of strings
+- Use datetime objects with UTC timezone for TIMESTAMP device class sensors
 
 ## [1.0.7]